+++ HISTORY.md
@@ -28,6 +28,7 @@
     * new "get graphic effect" reporter
     * new "get pen attribute" reporter
     * new "pen down?" predicate
+    * new "shown?" predicate
     * new "write" command in pen category (used to be "label" in tools)
     * new "numbers", "is empty", "map","keep", "find", "combine" and "for each" primitives in list category
     * new JIT-compiler "blitz-HOF" primitives for "map", "keep", "find" & "combine" via "compile"
@@ -86,6 +87,7 @@
 
 ### 2019-06-01
 * Objects: new "pen down?" predicate
+* Objects: new "shown?" predicate
 
 ### 2019-05-31
 * Threads: optimized FOREACH performance for large linked lists